Add compatibility check for community_maps
Allow for players to join and host vanilla games as normal and when a
community map is selected block users without mod with same version from
joining

Check communityMapsCompatibilityMatcher for the exact matching criteria.

Currently it will enable 0.0.11(.*) so if no changes to map generation
and it's safe for players with different versions to join append to
0.0.11, e.g. 0.0.11.1.

If compatibility is broken change mod version and compatiblility checker
to 0.0.12(.*) and so on.